requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions==1.4.0
pyyaml==5.3.1
snowflake-connector-python[pandas]==2.3.3
pandas==1.1.3
numpy==1.19.3
jinjasql==0.1.8